add llm rm old tera-model
