+++ packages/client/hmi-client/src/components/model/petrinet/model-configurations/tera-regular-model-configurations.vue
