+++ packages/client/hmi-client/src/components/model/petrinet/model-configurations/tera-stratified-model-configurations.vue
