+++ packages/client/hmi-client/src/components/model/petrinet/tera-model-editor.vue
@@ -23,7 +23,7 @@
 			/>
 		</AccordionTab>
 		<AccordionTab header="Model configurations">
-			<tera-model-configuration
+			<tera-model-configurations
 				:model="model"
 				:model-configurations="modelConfigurations"
 				:feature-config="featureConfig"
@@ -51,7 +51,7 @@ import { ref, onMounted, computed } from 'vue';
 import TeraModelDiagram from '@/components/model/petrinet/tera-model-diagram.vue';
 import TeraModelEquation from '@/components/model/petrinet/tera-model-equation.vue';
 import TeraModelObservable from '@/components/model/petrinet/tera-model-observable.vue';
-import TeraModelConfiguration from '@/components/model/petrinet/tera-model-configuration.vue';
+import TeraModelConfigurations from '@/components/model/petrinet/tera-model-configurations.vue';
 import { FeatureConfig, ResultType } from '@/types/common';
 import { Document, Dataset, Model, ProvenanceType, ModelConfiguration } from '@/types/Types';
 import Accordion from 'primevue/accordion';
