+++ packages/client/hmi-client/src/components/model/tera-model.vue
@@ -1,75 +1,82 @@
 <template>
-	<!--main is temporary just for css purposes-->
-	<main>
-		<tera-asset
-			v-if="model"
-			:name="model.header.name"
-			:feature-config="featureConfig"
-			:is-naming-asset="isNaming"
-			:stretch-content="view === ModelView.MODEL"
-		>
-			<template #name-input>
-				<InputText
-					v-if="isNaming"
-					v-model.lazy="newName"
-					placeholder="Title of new model"
-					@keyup.enter="updateModelName"
+	<tera-asset
+		v-if="model"
+		:name="model.header.name"
+		:feature-config="featureConfig"
+		:is-naming-asset="isNaming"
+		:stretch-content="view === ModelView.MODEL"
+		@close-preview="emit('close-preview')"
+	>
+		<template #name-input>
+			<InputText
+				v-if="isNaming"
+				v-model.lazy="newName"
+				placeholder="Title of new model"
+				@keyup.enter="updateModelName"
+			/>
+		</template>
+		<template #edit-buttons>
+			<span class="p-buttonset">
+				<Button
+					class="p-button-secondary p-button-sm"
+					label="Description"
+					icon="pi pi-list"
+					@click="view = ModelView.DESCRIPTION"
+					:active="view === ModelView.DESCRIPTION"
 				/>
+				<Button
+					class="p-button-secondary p-button-sm"
+					label="Model"
+					icon="pi pi-share-alt"
+					@click="view = ModelView.MODEL"
+					:active="view === ModelView.MODEL"
+				/>
+				<Button
+					class="p-button-secondary p-button-sm"
+					label="Transform"
+					icon="pi pi-sync"
+					@click="view = ModelView.NOTEBOOK"
+					:active="view === ModelView.NOTEBOOK"
+				/>
+			</span>
+			<template v-if="!featureConfig.isPreview">
+				<Button
+					icon="pi pi-ellipsis-v"
+					class="p-button-icon-only p-button-text p-button-rounded"
+					@click="toggleOptionsMenu"
+				/>
+				<Menu ref="optionsMenu" :model="optionsMenuItems" :popup="true" />
 			</template>
-			<template #edit-buttons>
-				<span class="p-buttonset">
-					<Button
-						class="p-button-secondary p-button-sm"
-						label="Description"
-						icon="pi pi-list"
-						@click="view = ModelView.DESCRIPTION"
-						:active="view === ModelView.DESCRIPTION"
-					/>
-					<Button
-						class="p-button-secondary p-button-sm"
-						label="Model"
-						icon="pi pi-share-alt"
-						@click="view = ModelView.MODEL"
-						:active="view === ModelView.MODEL"
-					/>
-					<Button
-						class="p-button-secondary p-button-sm"
-						label="Transform"
-						icon="pi pi-sync"
-						@click="view = ModelView.NOTEBOOK"
-						:active="view === ModelView.NOTEBOOK"
-					/>
-				</span>
-				<template v-if="!featureConfig.isPreview">
-					<!--Disable this until rename-dataset is merged-->
-					<Button
-						icon="pi pi-ellipsis-v"
-						class="p-button-icon-only p-button-text p-button-rounded"
-						@click="toggleOptionsMenu"
-					/>
-					<Menu ref="optionsMenu" :model="optionsMenuItems" :popup="true" />
-				</template>
-			</template>
-			<tera-model-description
-				v-if="view === ModelView.DESCRIPTION"
-				:model="model"
+		</template>
+		<tera-model-description
+			v-if="view === ModelView.DESCRIPTION"
+			:model="model"
+			:model-configurations="modelConfigurations"
+			:highlight="highlight"
+			:project="project"
+			@update-model="updateModelContent"
+			@fetch-model="fetchModel"
+		/>
+		<tera-model-editor
+			v-else-if="view === ModelView.MODEL"
+			:model="model"
+			:model-configurations="modelConfigurations"
+			:feature-config="featureConfig"
+			@update-model="updateModelContent"
+			@update-configuration="updateConfiguration"
+			@add-configuration="addConfiguration"
+		/>
+		<Suspense v-else-if="view === ModelView.NOTEBOOK">
+			<tera-model-jupyter-panel
+				:asset-id="props.assetId"
 				:model-configurations="modelConfigurations"
-				:highlight="highlight"
-				:project="project"
-				@update-model="updateModelContent"
-				@fetch-model="fetchModel"
-			/>
-			<tera-model-editor
-				v-else-if="view === ModelView.MODEL"
+				:project="props.project"
 				:model="model"
-				:model-configurations="modelConfigurations"
-				:feature-config="featureConfig"
-				@update-model="updateModelContent"
-				@update-configuration="updateConfiguration"
-				@add-configuration="addConfiguration"
+				:show-kernels="false"
+				:show-chat-thoughts="false"
 			/>
-		</tera-asset>
-	</main>
+		</Suspense>
+	</tera-asset>
 </template>
 
 <script setup lang="ts">
@@ -78,6 +85,7 @@ import { isEmpty, cloneDeep } from 'lodash';
 import TeraAsset from '@/components/asset/tera-asset.vue';
 import TeraModelDescription from '@/components/model/petrinet/tera-model-description.vue';
 import TeraModelEditor from '@/components/model/petrinet/tera-model-editor.vue';
+import TeraModelJupyterPanel from '@/components/model/tera-model-jupyter-panel.vue';
 import Button from 'primevue/button';
 import InputText from 'primevue/inputtext';
 import Menu from 'primevue/menu';
@@ -118,7 +126,11 @@ const props = defineProps({
 	}
 });
 
-const emit = defineEmits(['update-model-configuration', 'new-model-configuration']);
+const emit = defineEmits([
+	'close-preview',
+	'update-model-configuration',
+	'new-model-configuration'
+]);
 
 const model = ref<Model | null>(null);
 const modelConfigurations = ref<ModelConfiguration[]>([]);
@@ -225,16 +237,3 @@ watch(
 	{ immediate: true }
 );
 </script>
-
-<style scoped>
-/* main is temporary to mimic what tera-model is wrapped in */
-main {
-	display: grid;
-	grid-template-rows: auto 1fr;
-	height: 100%;
-	width: 100%;
-	background-color: var(--surface-section);
-	overflow-y: auto;
-	overflow-x: hidden;
-}
-</style>
